+++ src/app/projects/page.module.scss
@@ -1,88 +1,5 @@
 @use "../../app/variables.module.scss" as vars;
 
-.main {
-  display: flex;
-  height: calc(100vh - 72px);
-}
-
-.sidebar {
-  height: 100%;
-}
-
-.sidebarShow {
-  @extend .sidebar;
-  width: 360px;
-}
-
-.sidebarHidden {
-  @extend .sidebar;
-  width: 36px;
-}
-
-.sidebarContents {
-  height: calc(100% - 80px);
-}
-
-.sidebarHeader {
-  display: flex;
-  justify-content: flex-end;
-  height: 20px;
-  padding: 8px;
-}
-
-.sidebarToggle {
-  height: 0.8rem;
-  width: 0.8rem;
-  cursor: pointer;
-}
-
-.sidebarBody {
-  height: 100%;
-}
-
-.sidebarBodyList {
-  list-style-type: none;
-  padding-left: 0%;
-}
-
-.sidebarBodyListItem {
-  display: block;
-  padding: 16px 16px 16px 32px;
-  font-size: 12px;
-  white-space: nowrap;
-  &:hover {
-    background-color: vars.$sidebarHoverBgColor;
-    cursor: pointer;
-  }
-}
-
-.sidebarBodyListItemTitle {
-  color: vars.$fontColorDark;
-  font-weight: 500;
-  letter-spacing: 0.8px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.sidebarProjectListItem {
-  display: flex;
-  justify-content: space-between;
-  padding: 8px 16px 8px 16px;
-  font-size: 10px;
-  white-space: nowrap;
-  &:hover {
-    background-color: vars.$sidebarHoverBgColor;
-    cursor: pointer;
-  }
-}
-
-.sidebarProjectDeadline {
-  display: inline-block;
-  font-size: 8px;
-  margin-left: 32px;
-}
-
 .mainContents {
   padding: 32px;
   width: 100%;
